# docs/swagger.yaml
basePath: /
definitions:
...
  handler.createRoomRequest:
    properties:
      capacity:
        example: 8
        type: integer
      description:
        example: Большая переговорка
        type: string
      name:
        example: Переговорка 1
        type: string
    type: object
  handler.createScheduleRequest:
    properties:
      daysOfWeek:
        example:
        - 1
        - 2
        - 3
        - 4
        - 5
        items:
          type: integer
        type: array
      endTime:
        example: "18:00"
        type: string
      roomId:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      startTime:
        example: "09:00"
        type: string
    type: object
...
  handler.roomResponse:
    properties:
      room:
        $ref: '#/definitions/model.Room'
    type: object
  handler.roomsResponse:
    properties:
      rooms:
        items:
          $ref: '#/definitions/model.Room'
        type: array
    type: object
  handler.scheduleResponse:
    properties:
      schedule:
        $ref: '#/definitions/model.Schedule'
    type: object
  handler.slotsResponse:
    properties:
      slots:
        items:
          $ref: '#/definitions/model.Slot'
        type: array
    type: object
...
  httputil.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/httputil.errorBody'
    type: object
  httputil.errorBody:
    properties:
      code:
        example: INVALID_REQUEST
        type: string
      message:
        example: invalid request
        type: string
    type: object
...
  model.Room:
    properties:
      capacity:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  model.Schedule:
    properties:
      daysOfWeek:
        items:
          type: integer
        type: array
      endTime:
        type: string
      id:
        type: string
      roomId:
        type: string
      startTime:
        type: string
    type: object
  model.Slot:
    properties:
      end:
        type: string
      id:
        type: string
      roomId:
        type: string
      start:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Сервис бронирования переговорок
  title: Room Booking Service
  version: "1.0"
paths:
...
  /rooms/{roomId}/schedule/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: ID переговорки
        format: uuid
        in: path
        name: roomId
        required: true
        type: string
      - description: Расписание
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createScheduleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.scheduleResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать расписание переговорки
      tags:
      - Schedules
  /rooms/{roomId}/slots/list:
    get:
      parameters:
      - description: ID переговорки
        format: uuid
        in: path
        name: roomId
        required: true
        type: string
      - description: Дата (YYYY-MM-DD)
        example: "2024-06-10"
        in: query
        name: date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.slotsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список доступных слотов по переговорке и дате
      tags:
      - Slots
  /rooms/create:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные переговорки
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.createRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.roomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать переговорку
      tags:
      - Rooms
  /rooms/list:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.roomsResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список переговорок
      tags:
      - Rooms
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"